feat: format target description in search results
