+++ src/pages/TargetPage/ProfileHeader.js
@@ -6,6 +6,7 @@ import {
   ChipList,
 } from '../../components/ProfileHeader';
 import usePlatformApi from '../../hooks/usePlatformApi';
+import { clearDescriptionCodes } from '../../utils/global';
 
 const TARGET_PROFILE_HEADER_FRAGMENT = loader('./TargetProfileHeader.gql');
 
@@ -54,30 +55,15 @@ const parseSynonyms = synonyms => {
   return parsedSynonyms;
 };
 
-// TODO: Replace this with PublicationsDrawer component
-function makePmidLink(match) {
-  const id = match.substring(7);
-  const linkStyles =
-    'color: #3489ca; font-size: inherit; text-decoration: none;';
-  return `PubMed:<a style="${linkStyles}" href="https://europepmc.org/abstract/med/${id}" target="_blank">${id}</a>`;
-}
-
-function clearCodes(descriptions) {
-  if (!descriptions) return [];
-  return descriptions.map(desc => {
-    const codeStart = desc.indexOf('{');
-    const parsedDesc = desc.slice(0, codeStart);
-    return parsedDesc.replace(/Pubmed:\d+/gi, makePmidLink);
-  });
-}
-
 function ProfileHeader() {
   const { loading, error, data } = usePlatformApi();
 
   //TODO: Errors!
   if (error) return null;
 
-  const targetDescription = clearCodes(data?.target.functionDescriptions);
+  const targetDescription = clearDescriptionCodes(
+    data?.target.functionDescriptions
+  );
   const synonyms = parseSynonyms(data?.target.synonyms || []);
 
   return (
