+++ src/utils/global.js
@@ -40,3 +40,20 @@ export function getUniprotIds(proteinIds) {
     .filter(proteinId => proteinId.source === 'uniprot_swissprot')
     .map(proteinId => proteinId.id);
 }
+
+// TODO: Replace this with PublicationsDrawer component
+export function makePmidLink(match) {
+  const id = match.substring(7);
+  const linkStyles =
+    'color: #3489ca; font-size: inherit; text-decoration: none;';
+  return `PubMed:<a style="${linkStyles}" href="https://europepmc.org/abstract/med/${id}" target="_blank">${id}</a>`;
+}
+
+export function clearDescriptionCodes(descriptions) {
+  if (!descriptions) return [];
+  return descriptions.map(desc => {
+    const codeStart = desc.indexOf('{');
+    const parsedDesc = desc.slice(0, codeStart);
+    return parsedDesc.replace(/Pubmed:\d+/gi, makePmidLink);
+  });
+}
